rename sub-command-files
